:recycle: Fix cryptography tests
